Feature/submission-crud: Refactor answer folder and constructors
