+++ src/main/java/com/tophat/teacherdemo/controller/ProblemController.java
@@ -1,10 +1,10 @@
 package com.tophat.teacherdemo.controller;
 
 import com.tophat.teacherdemo.controller.vo.ProblemCreateRequest;
-import com.tophat.teacherdemo.entity.Answer;
+import com.tophat.teacherdemo.entity.answer.Answer;
 import com.tophat.teacherdemo.entity.Problem;
 import com.tophat.teacherdemo.service.ProblemService;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,10 +13,11 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/teacherdemo/v0/problem")
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class ProblemController {
     private final ProblemService problemService;
 
+    // TODO: Restrict API access to prevent answer leaks
     @GetMapping("/{id}")
     public ResponseEntity<Problem<Answer>> getProblem(@PathVariable ObjectId id) {
         Optional<Problem<Answer>> foundAnswer = problemService.findProblemById(id);
