+++ src/main/java/com/tophat/teacherdemo/entity/answer/Answer.java
@@ -1,4 +1,4 @@
-package com.tophat.teacherdemo.entity;
+package com.tophat.teacherdemo.entity.answer;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
