+++ src/main/java/com/tophat/teacherdemo/entity/answer/MultipleChoiceAnswer.java
@@ -1,4 +1,4 @@
-package com.tophat.teacherdemo.entity;
+package com.tophat.teacherdemo.entity.answer;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import lombok.Data;
