+++ src/main/java/com/tophat/teacherdemo/repository/ProblemRepository.java
@@ -1,6 +1,6 @@
 package com.tophat.teacherdemo.repository;
 
-import com.tophat.teacherdemo.entity.Answer;
+import com.tophat.teacherdemo.entity.answer.Answer;
 import com.tophat.teacherdemo.entity.Problem;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
